% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.tip.random.r
\encoding{UTF-8}
\name{add.tip.random}
\alias{add.tip.random}
\title{Add tips at random to the tree}
\usage{
add.tip.random(tree, tip, method = "length")
}
\arguments{
\item{tree}{phylogeny as an object of class "phylo".}

\item{tip}{tip name for the added.}

\item{method}{The method to add the new tip, partial match to "equal" or "length". See details (Default method = "length").}
}
\value{
The phylogeny as an object of class "phylo" with new tip.
}
\description{
Function to add new tip at random to a tree with branch length. This function is a small change in the function \code{\link{add.random}}, see details.
}
\details{
This function is a small change in the function \code{\link{add.random}} allowing add only one tip at random only in ultrametric phylogenetic tree. 
Two methods are available. The method "length" add the tip with probability equivalent to edge length, long edge have more probability than short edge. 
The method "equal" each edge has equal probability to receive the new tip.
}
\examples{
set.seed(10)
tree <- rtree(5)
tree <- compute.brlen(tree)
tree <- makeNodeLabel(tree)
plot.phylo(tree, show.node.label = TRUE)
res <- add.tip.random(tree, "NEW", method = "length")
res
plot.phylo(res, show.node.label = TRUE)
}
\seealso{
\code{\link{add.random}} \code{\link{add.taxa.phylo}}
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{daee}
