% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact.tree.R
\encoding{UTF-8}
\name{compact.tree}
\alias{compact.tree}
\title{Compact a phylogenetic tree}
\usage{
compact.tree(tree, nodes)
}
\arguments{
\item{tree}{phylogeny as an object of class "phylo".}

\item{nodes}{a vector with node label to compact the edges. Only two tips are keeps, most short a most long within each nodes}
}
\value{
The phylogeny as an object of class "phylo" with nodes compacted.
}
\description{
Funtion to compact a phylogenetic tree. See details.
}
\details{
This function compacts the node of a phylogenetic tree. Inside each node the function removes tips
using the function \code{\link{drop.tip}}. Only two tips are keeped: the longhest and shortest edge length within each 
node. This can be useful only for the function \code{\link{plotcollapse.phylo}}.
}
\examples{
set.seed(10)
tree <- rtree(15)
tree <- makeNodeLabel(tree)
plot.phylo(tree, show.node.label = TRUE)
nodes <- c("Node6", "Node13")
res <- compact.tree(tree, nodes)
res
plot.phylo(res, show.node.label = TRUE)
}
\seealso{
\code{\link{plotcollapse.phylo}} \code{\link{graphical.node.patterns}}
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{daee}
