\name{daee-package}
\alias{daee-package}
\alias{daee}
\docType{package}
\title{
  Data Analysis for Ecology and Evolution
}
\description{
  Set of functions for analyzing ecological and evolutionary data. The functions are available as an R package only to facilitate installation.
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@@yahoo.com.br>

Maintainer: Vanderlei Julio Debastiani <vanderleidebastiani@@yahoo.com.br>
}
\keyword{daee}