% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantel.residuals.R
\encoding{UTF-8}
\name{mantel.residuals}
\alias{mantel.residuals}
\title{Residuals from Mantel test}
\usage{
mantel.residuals(y, x)
}
\arguments{
\item{y}{Dissimilarity matrices of dist class.}

\item{x}{Dissimilarity matrices of dist class.}
}
\value{
\item{statistic}{The Mantel statistic.} \item{residuals}{Residuals 
extracted from the Mantel test (Class dist).}
}
\description{
Function to extract residuals from Mantel test.
}
\examples{

#require(vegan)
#A<-vegdist(matrix(sample(1:10, 20, replace  =TRUE), 5, 5))
#B<-vegdist(matrix(sample(1:10, 30, replace = TRUE), 5, 6))
#mantel(A, B)
#mantel.residuals(A, B)

}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{daee}

