% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.tree.R
\encoding{UTF-8}
\name{node.tree}
\alias{node.tree}
\title{Makes node labels}
\usage{
node.tree(tree, m = 0, prefix = "NodE")
}
\arguments{
\item{tree}{phylogeny as an object of class "phylo".}

\item{m}{One number to starting sequence (default m = 0).}

\item{prefix}{The prefix (default prefix = "NodE")}
}
\value{
A list with: \item{call}{The call arguments.}\item{m.start}{The start m value.} \item{m.current}{The new m value, the difference between m.start and m.current is the number of node names created.} \item{prefix}{Prefix used.} \item{tree}{The tree, class "phylo".}
}
\description{
This function makes node labels in a tree keeping already existing node names. This function is based in the function \code{\link{makeNodeLabel}}.
}
\examples{
tree <- read.tree(text = "(C:32,(B:16,A:16):16);")
plot(tree, show.node = TRUE)
tree <- node.tree(tree, m = 0, prefix = "N")$tree
plot(tree, show.node = TRUE)
}
\seealso{
\code{\link{makeNodeLabel}}
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{daee}
