% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.list.R
\encoding{UTF-8}
\name{organize.list}
\alias{organize.list}
\title{Organize a list with data.frame/matrix in a single matrix}
\usage{
organize.list(x, force.names = FALSE)
}
\arguments{
\item{x}{A list with data.frame or matrix}

\item{force.names}{Logical argument (TRUE or FALSE) to specify if new row names are created, using list name as prefix.}
}
\value{
A matrix combined.
}
\description{
Function organize a list of data.frame or matrix in a single matrix. The function combine the data.frame/matrix by
columns, thus all objects in the list must be equal number of columns
}
\examples{
A<-matrix(1,10,6)
rownames(A) <- rownames(A, do.NULL = FALSE, prefix = "A.")
B<-matrix(2,3,6)
rownames(B) <- rownames(B, do.NULL = FALSE, prefix = "B.")
C<-matrix(3,15,6)
rownames(C) <- rownames(C, do.NULL = FALSE, prefix = "C.")
RES<-list(as.data.frame(A), as.data.frame(B), as.data.frame(C))
organize.list(RES)
organize.list(RES, force.names = TRUE)
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{daee}
