% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVR.adonis.R
\encoding{UTF-8}
\name{PVR.adonis}
\alias{PVR.adonis}
\title{Phylogenetic Eigenvector Regression (PVR) and eigenvector selection}
\usage{
PVR.adonis(traits, dist, cumulative = 0.99)
}
\arguments{
\item{traits}{Data matrix or a dissimilarity matrix (recommended), usually related to species traits. 
This will be passed to the left side of the formula in the adonis function. The sequence of species in 
the traits data matrix or dissimilarity matrix must be the same as that in the phylogenetic distance 
matrix. See details in \code{\link{adonis}}.}

\item{dist}{Phylogenetic distance matrix.}

\item{cumulative}{Percentage of variation in the phylogenetic distances 
considered in the analysis. Cumulative percentage must be higher than the 
cumulative percentage of the first two eigenvalues, and less than 1.}
}
\value{
\item{values}{Eigenvalues, relative eigenvalues and cumulative eigenvalues 
for the PCoA of distance matrix.} \item{vectors}{The principal coordinates with 
positive eigenvalues.} \item{inf.cumulative}{Percentage of the variation in the 
phylogenetic distances considered in the analysis (The result should be approximately
the specified cumulative value).} \item{n.axis.considered}{Number of axes considered.}
\item{results.unique}{F value for each PVR axis} \item{results.sequential}{F value for 
sequential approach using all PVR axes (PVR 1,PVR 1 + PVR 2, ...).} 
\item{results.stepwise}{F value for non-sequential approach, that uses the combination 
of PVRs axes that maximize the F value. The selection finishes using all PVRs 
considered, but the max F value must be selected manually in the results.}
}
\description{
Phylogenetic Eigenvector Regression (PVR) and eigenvector selection 
using analysis of variance with distance matrices (adonis).
}
\details{
The phylogenetic distance matrix is double-centered and submitted to principal
coordinates analysis (PCoA). This method generates orthogonal eigenvectors that
summarize the phylogenetic structure (Diniz-Filho et al 2008).

This function is similar the function \code{\link{PVR}} that use a non-sequential approach 
to perform the eigenvector selection, but the selection is based in multivariate
analysis of variance. The function search to combination of eigenvectors that maximize 
the F value in the analysis of variance with distance matrices using the \code{\link{adonis}} function. 
Primarily, an analysis for each eigenvectors is performed, obtaining the F values. Then, the function 
select the eingenvector with the higher F value, and then, new eigenvectors are added to the model, 
models are updated and F values are obtained. The search stops when all eigenvectors are included in 
the model. The subset of eigenvectors that maximize the global F value must be selected manually in the results.
}
\examples{

# require(SYNCSA)
# require(vegan)
# data(flona)
# traits.dist <- vegdist(decostand(flona$traits[,c(1,3)], 
# 								 method = "standardize"), 
# 					   method = "euclidean")
# results <- PVR.adonis(traits.dist, flona$phylo, cumulative = 0.7)
# results
# plot(factor(results$results.unique$PVR, levels =results$results.unique$PVR), 
# 	 results$results.unique$F.value, 
# 	 xlab = "PVR", ylab = "F value", main = "results.unique")
# plot(factor(results$results.sequential$PVRs, levels = results$results.sequential$PVRs), 
# 	 results$results.sequential$F.value,
# 	 xlab = "PVRs", ylab = "F value", main = "results.sequential")
# plot(factor(results$results.stepwise$PVRs, levels = results$results.stepwise$PVRs), 
# 	 results$results.stepwise$F.value,
# 	 xlab = "PVRs", ylab = "F value", main = "results.stepwise")

}
\references{
Diniz-Filho, J. A. F., Santana, C. E. R., Bini, L. M. (1998). An 
eigenvector method for estimating phylogenetic inertia. Evolution, 52(5), 1247-1262.
}
\seealso{
\code{\link{PVR}}
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{daee}
