% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.label.info.R
\encoding{UTF-8}
\name{tree.label.info}
\alias{tree.label.info}
\title{Show information about a label in a phylogenetic tree}
\usage{
tree.label.info(tree, label)
}
\arguments{
\item{tree}{phylogeny as an object of class "phylo".}

\item{label}{One tip label or one node label.}
}
\value{
A data.frame with type of label, edge position, edge length, edge height and max height of tree.
}
\description{
Function for discover the edge position, edge length, edge height, max height of tree and type of the
label in a phylogenetic tree. This function is based in the function \code{\link{nodeHeights}}.
}
\examples{
tree <- read.tree(text = "(C:32,(B:16,A:16)N2:16)N1;")
plot(tree, show.node = TRUE)
tree.label.info(tree, label = "A")
}
\seealso{
\code{\link{nodeHeights}}
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{daee}
